# doc/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mavka Open API Specification",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Dev server"
    }
  ],
...
  "paths": {
...
    "/groups/{group_id}": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Join a group",
        "description": "This method is for joining groups",
        "operationId": "joinGroup",
        "tags": ["Groups"],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group id to join",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Success"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/news": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get news",
        "description": "This method returns recomended posts",
        "operationId": "getNews",
        "tags": ["Posts"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "post_id": {
                        "type": "integer"
                      },
                      "username": {
                        "type": "string"
                      },
                      "avatar_url": {
                        "type": "string"
                      },
                      "wall_id": {
                        "type": "integer"
                      },
                      "post_text": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "text"
                      }
                    }
                  },
                  "example": [
                    {
                      "post_id": 6,
                      "username": "gervant",
                      "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg",
                      "wall_id": 6,
                      "post_text": "Хэллоу ворлд!",
                      "created_at": "2024-01-19T11:24:44.886Z"
                    },
                    {
                      "post_id": 5,
                      "username": "gervant",
                      "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg",
                      "wall_id": 6,
                      "post_text": "NEW POST TEXT",
                      "created_at": "2024-01-19T10:50:26.441Z"
                    },
                    {
                      "post_id": 4,
                      "username": "gervant",
                      "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg",
                      "wall_id": 6,
                      "post_text": "Я поймал и расколдовал стрыгу!!! :D",
                      "created_at": "2024-01-19T10:50:26.441Z"
                    },
                    {
                      "post_id": 3,
                      "username": "gervant",
                      "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg",
                      "wall_id": 6,
                      "post_text": "Я поймал и расколдовал стрыгу!!! :D",
                      "created_at": "2024-01-19T10:50:26.441Z"
                    },
                    {
                      "post_id": 2,
                      "username": "gervant",
                      "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg",
                      "wall_id": 6,
                      "post_text": "Я поймал и расколдовал стрыгу!!! :D",
                      "created_at": "2024-01-19T10:50:26.441Z"
                    },
                    {
                      "post_id": 1,
                      "username": "gervant",
                      "avatar_url": "https://i.ibb.co/t33h8S6/witcher.jpg",
                      "wall_id": 6,
                      "post_text": "Я поймал и расколдовал стрыгу!!! :D",
                      "created_at": "2024-01-19T10:50:26.441Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}